refactor: retorno de average rating é double

+++ src/main/java/rocha/andre/api/domain/fullGame/DTO/FullGameUserDTO.java
@@ -7,9 +7,9 @@ import java.util.UUID;
 
 public record FullGameUserDTO(UUID id, String name, Integer metacritic, Integer yearOfRelease,
                                 ArrayList<String> studios, ArrayList<String> genres, ArrayList<String> consoles, String imageUrl,
-                                int totalReviews, int averageRating) {
+                                int totalReviews, double averageRating) {
 
-    public FullGameUserDTO(Game game, ArrayList<String> studios, ArrayList<String> genres, ArrayList<String> consoles, String imageUrl, int totalReviews, int averageRating) {
+    public FullGameUserDTO(Game game, ArrayList<String> studios, ArrayList<String> genres, ArrayList<String> consoles, String imageUrl, int totalReviews, double averageRating) {
         this(game.getId(), game.getName(), game.getMetacritic(), game.getYearOfRelease(), studios, genres, consoles, imageUrl, totalReviews, averageRating);
     }
 }
