+++ src/main/java/rocha/andre/api/domain/gameRating/DTO/AllRatingsGameDTO.java
@@ -1,4 +1,4 @@
 package rocha.andre.api.domain.gameRating.DTO;
 
-public record AllRatingsGameDTO(int totalReviews, int averageRating) {
+public record AllRatingsGameDTO(int totalReviews, double averageRating) {
 }
